+++ src/library/DIPS.Mobile.UI/Components/Loading/StateView/IStateChangedAware.cs
@@ -1,6 +1,0 @@
-namespace DIPS.Mobile.UI.Components.Loading.StateView;
-
-public interface IStateChangedAware
-{
-    Task OnStateChanged(State state);
-}